Beta 6.3

+++ backend/gesmerca/database/migrations/2023_03_20_110122_create_product.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('supplier');
             $table->foreign('supplier')->references('id')->on('suppliers')->onDelete('cascade');
             $table->text('image');
-            $table->double('price');
+            $table->float('price');
             $table->integer('stock');
             $table->timestamps();
         });
